pyproject.toml
[build-system]
requires = [
  "poetry-core>=1.0.0",
  "setuptools>=42",
  "wheel",
]
# build-backend = "setuptools.build_meta"
build-backend = "poetry.core.masonry.api"

[tool.poetry]
authors = ["Yicheng Luo <yicheng.luo.20@ucl.ac.uk>", "Antonio Filieri <a.filieri@imperial.ac.uk>"]
description = "SYMPAIS: Symbolic Parallel Adaptive Importance Sampling for Probabilistic Program Analysis"
name = "sympais"
version = "0.0.1"
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
hydra-core = "*"
hydra-joblib-launcher = "*"
jax = "<=0.2.13"
jaxlib = "<=0.1.65"
numpy = "*"
numpyro = "<=0.6.0"
pandas = "*"
ply = "<=3.11"
python = ">=3.7, <3.10"
scipy = "*"
sympy = "<=1.8"
z3-solver = "<=4.8.10.0"
PySMT = "^0.9.0"

[tool.poetry.dev-dependencies]
isort = "*"
pylint = "*"
pytest = "*"
pytest-xdist = "*"
pytype = ">=2021.5.25"
yapf = "*"
pre-commit = "^2.13.0"
flake8 = "^3.9.2"

[tool.isort]
profile = "google"
src_paths = ["src"]
